DEV-3448 Fix aws-sdk-cpp build on latest msvc compiler
Due to the msvc compiler change in https://github.com/microsoft/STL/pull/3818,
the aws-sdk-cpp CI build breaks (warnings are treated as errors).

A fix is available upstream when we upgrade our vcpkg ports but to avoid
an immediate upgrade, we'll define _SILENCE_STDEXT_ARR_ITERS_DEPRECATION_WARNING
for now.